README: Remove the data volume container approach.
Even though this seems to be the recommended approach according to
https://docs.docker.com/userguide/dockervolumes/ the support for managing the
data in the volumes is still lacking. Snapshotting the container for example
does not capture the data stored in the volume hence it is not a suitable
approach to create backups yet.

By removing the data volume approach we can remove the data volume image itself
(less maintenance) and we only talk about a single approach.

This is a reversal of the decision made in
https://bitbucket.org/atlassian/docker-atlassian-stash/pull-request/1/first-cut-of-an-official-stash-dockerfile