Add ssh to installed binaries
Currently there is no ssh binary on the path for the docker image. This
is fine for running the main image but will cause failures if the image
is run in mirroring mode. This is due to the fact that the main
application forms its SSH connections via Apache MINA, therefore with no
need for the ssh binary. Mirroring on the other hand is forking to the
git process to perform git clone and git fetch.

This means that in the event that `git clone ssh://some-original-url` is
called, the call will fail as git will interact with an ssh binary that
it expects to exist.